Adds BlindMarshaler / BlindUnmarshaler
Also removes redundant Marshaler / Unmarshaler checks within the
individual methods, as those cases should be covered by the overall
Marshal / Unmarshal methods.

There is an edge case that falls through but I think it's wrong
behavior: the old code would use the Marshaler interface in the case
where the value is a pointer to a Marshaler but is not itself a
Marshaler (this comes from the recursive calls in the reflect case).

Since a pointer to a struct implements the same interfaces as the
underlying struct, this only comes up when it's a pointer to a pointer
to a struct, which you shouldn't be trying to marshal anyway.