Some inline code comments for beginners
The program panics when the keyspace or table was not available. The querying
part failed when an index was not prebuilt.

These information may be common knowledge for Cassandra experts. But I was a
beginner who was considering cassandra and gocql over sqlite and had to spend
time finding these pre-requisite cql instructions myself. Since it is the basic
example, Documenting these CQL queries will help beginners.