Refactor recv loop, inline dispatch
Make the recv loop simpler, pass the error back directly from
recv so that we dont close the connection on non fatal errors.

Use a mutex to guard the read/write buffer inside the frame from
the connection, it appears we can start reading a frame into the
framers buffer as soon as we have finished writing the frame out.