refactor TypeInfo into an interface
Make TypeInfo an interface in preperation for implementing
marshalling logic for Tuples and potentially UDT's and other
types which Cassandra might add in the future.

The previous implementation was rigid in that it assumed a single
returned column for a returned type, but with tuples and UDT's this
is no longer the case as the tuple will have a single TypeInfo for
multiple columns.

This change will break existing clients which implement MarshalCQL
or UnmarshalCQL as the TypeInfo is passed in is not a pointer so the
interface type assertion will not match the expected interface.