Defensivly copy byte slices from the read buffer
When parsing frames which contains [bytes] copy the underlying
bytes into a new slice so that the owner as full control of the
lifecycle of the slice. This has the downside that it increases
our allocations by about a factor of 2 in the benchmarks but it
overcomplicates the assumptions that need to be made about the
underlying memory which will more likely than not lead to a subtle
memory corruption bug.