Dont hold the write lock when updating hosts
In the PolicyAwareConnection pool we shouldnt hold the write lock
whilst dialing hosts as it will block all readers. In the event that
a host is down we will try to connect to it and wait until timeout
before stopping. This means that whilst the connection is dialing
all readers will be blocked acquiring the read lock to access the
pools.

Instead change the policy aware pools to use a copy on write map
which allows us to do the mutations to the map without having the
write lock.