apply a timeout for every query at the connection
Start a timeout when waiting for the response from cassandra
in the call to exec() so that if no header is read the query
will not hang for ever.

If there are more than a limit of timeouts then the connection will
be closed. This may need tuning in the future if there are queries
which take a long time.

fixes #416