Marshal fixed for pointers implementing Marshaler
This is a fix for the following situation:

  type T struct { /* ... */ }

  // MarshalCQL implements the gocql.Marshaler interface
  func (t *T) MarshalCQL(info TypeInfo) ([]byte, error) {
      // ...
  }

  func laterInTheCode() {

      t := &T{ /* ... */ }

      data, err := gocql.Marshal(info, t) // <-- won't work

      // ...
  }

Marshal resolves a pointer value before checking if it implements the
Marshaler interface. The code above won’t work because T doesn’t
implement the Marshaler interface, only *T does.

I added a test that shows the bug. The test fails without the fix and
passes with it.