conn: ensure that callers unblock correctly
Make the response channel unbuffered so that there is an
explicit synchronisation point between exec() and recv(), this
prevents a case where a timeout could happend and recv() will still
send a response to the waiting response channel even though the caller
is not receving.

In closeWithError if the error is nil dont send it to waiting callers
because they will try to access the response frame, instead close
the quit channel which will indicate the connection has been closed
and should no longer wait for a response.