Improve support for complicated cassandra schemas.
In some schemas there may be "partition_key" type columns returned in the V2
protocol metadata which have the same component index; in our case we have
a couple column families with aliased columns "KEY" and "key" for historical
reasons. This change is based on the DataStax driver's behavior for columns
with the same component index, which prefers the last column in order defined
for that component index; the consequence of this behavior is that other column
names defined for the same component index will not be identified as part of a
routing key.

Also, "compact_value" has been added as a column kind/type.