conn: ensure that call.framer is not nil
If a connection is closed then call to closeWithError will send
the error to the callers resp channel, which the caller in exec
will unblock, release the stream which will set the framer to nil.
Then if the response for that stream comes back from Cassandra into
recv() then the call.framer will be nil and will panic.

Instead of always releasing the stream when getting a reponse from
the channel only do it if there is no error.

Updates #439