Don't crash on bad peers.
It's possible (or at least was possible), to be left with invalid
system.peers entries in your db after changing your
ring (https://issues.apache.org/jira/browse/CASSANDRA-6053 is at least
one way it used to be able to happen). If this happens (or someone who
is already in this state upgrades gocql and is trying to doing peer
discovery), ideally gocql would continue working on the good peers
instead of crashing trying to parse the bad peers. I made two changes
with this goal in mind:

1. Fix version string parsing to not crash if version is too short.
2. Skip peers if we fail parsing the version string (but still return
   the peers we were able to parse).
3. When waiting for schema agreement, skip peers that have an empty
   schema_version (otherwise it times out and fails).

Our ring had peer entries with null "rack", "release_version",
"schema_version", and "tokens".