close call.timeout to prevent deadlock. fixes #664
if a successful response is recieve at the same time as the 10th
timeout, a deadlock can occur if succesfull response is read
from the response channel before closeWithError send an error
of response channel. By closing the call.timeout after reading
from the response channel we can avoid this deadlock.