conn: pass context through startup chain
The startup procoss is very concurrent, and can fail in different
places. To ensure we dont end up blocking somewhere indefinetly make use
a context.

Run the startup loop concurrently whilst waiting for an error from recv
so that we get the recv error instead of just timing out waiting for
startup to finish.