marshall: support small and tiny ints, fix uint
Handle cql small int and tiny int, these are 16 bit and 8 bit signed
ints.

Fix unmarshalling uints to correctly set the max value of the int when
the sign bit is extended to a larger type. This fixes it so that it
correctly unmarshalls a tiny int of value 0xFF into any uint 0xFF
instead of the max value of the unit type.