Implemented WithTimestamp function in gocql.Batch
This function is an addition to DefaultTimestamp function.
Using this function enables defaultTimestamp flag, but instead of
generating new timestamp, it is possible to set it's value explicitly.

It's equal to USING TIMESTAMP x in the statement, but does not break
prepared query optimization.

Similar to #706 PR but for batches.