adding AddressTranslator interface and impl for use in ec2
This change introduces the AddressTranslator interface, which is
intended to translate peer addresses just before creating a connection
to those nodes. The primary use -- which is driving the change -- is
to be able to translate public IPs to private IPs in ec2.

This solution is common among other CQL driver implementations. The
specific implementation here also follows the convention set by
HostFilter.

Signed-off-by: Justin "Gus" Knowlden <gus@gusg.us>