change HostInfo.Peer to be an IP
When Cassandra returns us a hosts info the peer address is defined as an
inet both at protocol for events and the schema for the peer
information.

Previously we stored this as a string, and used it to
connect to hosts and also to index hosts by. This is different to what
use for user supplied address endpoints, we keep the potentially DNS
name as the peer address. This means that we can end up having duplicate
host pools, duplicate host info in the ring.

Fix this by making everything rely on a hosts address being an IP
address instead of either a DNS name or an IP.