marshal: Add expectations to errors for Marshaling
What
===
Add explicit expectations to exactly what error is returned in
test cases where the marshal and unmarshal results in an error.

Why
===
In cases where an unmarshal fails the presence of an error was expected
but not what error was returned. Marshal is not tested for the cases
where it fails in the table tests. Both of these are needed to add
more tests that test for new errors being returned.