marshal: Add an error when UUID has wrong length
What
===
Add a descriptive error when a UUID is attempted to be marshaled or
unmarshaled and the length of the []byte is not exactly 16 bytes long.

Example
===

```
can not marshal []byte 6 bytes long into timeuuid, must be exactly 16 bytes long
```

Why
===
In cases where UUID []byte was not the correct length the generic `can
not marshal %T into %s` error was being returned. The error was
misleading because it gave the impression the root cause was the type of
the data and not it's length. Additionally the `%T` formatter for a
`[]byte` prints out `[]uint8` which was even more confusing.