SslOptions: embed a pointer to the TLS config (#891)
* SslOptions: embed a pointer to the TLS config

This embeds a pointer to the desired TLS config instead of a concrete
value. Because Go copies the arguments of all function calls, this has the side effect of copying the TLS configuration, which copies a mutex. This causes
`go vet` to report that `crypto/tls.Config contains sync.Once contains sync.Mutex`.

This patch removes this complaint, but at the cost of needing people to change
their cassandra TLS configuration code to remove the dereference of the TLS
config.

* connectionpool: fix build

* setupTLSConfig: create TLS config if it does not exist