marshal: fix is nullable check
When we read null from the wire it is encoded as -1 which gets returned
from the frame as []byte(nil), this is distinct from the framer reading
a length 0 byte which is returned as []byte{}, these both have 0 length
but one is nil the other is empty.

Fix marshalling to correctly handle nil and empty values.