Rework attempt timeout implementation
This replaces the attempt timeout via `context.WithDeadline` with a
cancelable context and a recycled timer on the query akin to the
strategy used in #661. The unit test fix is actually a side effect
of the overhead saved by recycling the timer. Previously each attempt
took ~27ms (on a 25ms attempt timeout) because of the overhead of
`context.WithDeadline`. It's now more like 25.2ms which occasionally
resulted in a fifth query attempt being made because the 100ms query
context deadline timer fired a millisecond later.