Fix bug in query executor that prevented retries from ceasing (#1164)
If a retry policy is defined for the session/query, retries wouldn't
stop even if the retry policy returned false in the call to `Attempt()`.
The bug was introduced in #1151. As long as queries keep failing,
the query would be retried. The unit test that checks the retries
being executed now checks for an exact number of query retries, making
it a regression test as well.