reduce number of gourtines used to execute queries
When a query is idempotent or speculation is disabled do not launch a
new goroutine.

When we are speculating don't launch a goroutine to do the speculation
instead do the check in line for results while waiting for a speculation
tick.

This also removes the need to spawn a ticker when speculation is
disabled or the query is non idempotent.