repair write coalescing
Current write coalescing depends on the idea that the passed in
connection can do the writev syscall.

In Go only net.Conns implemented by the net package have this property,
since that the mechanism requires an implementation of a private method,
checked against an interface private to the net package.

So in practise only real TCP connections implement that.

TLS connections cannot do that either, but that is not handled here.

To get a real tcp connection, I broke up the abstraction of the
deadlineWriter and put the effect of it into the flush call.

To make it clearer that we really need a network connection, I changed
the interface passed in to the clearer net.Conn.

This is just the minimal fix to get writev back. A little cleanup work
is needed to.

Without this fix, we get only delayed writes which later happen
in quick succession instead of coalesced writes,
which adds only latency without any beneficial effect.

Coalescing tests needed adaption, since we really need a net.Conn now.