repair write coalescing (#1245)
* repair write coalescing

Current write coalescing depends on the idea that the passed in
connection can do the writev syscall.

In Go only net.Conns implemented by the net package have this property,
since that the mechanism requires an implementation of a private method,
checked against an interface private to the net package.

So in practise only real TCP connections implement that.

TLS connections cannot do that either, but that is not handled here.

To get a real tcp connection, I broke up the abstraction of the
deadlineWriter and put the effect of it into the flush call.

To make it clearer that we really need a network connection, I changed
the interface passed in to the clearer net.Conn.

This is just the minimal fix to get writev back. A little cleanup work
is needed to.

Without this fix, we get only delayed writes which later happen
in quick succession instead of coalesced writes,
which adds only latency without any beneficial effect.

Coalescing tests needed adaption, since we really need a net.Conn now.

* tls and write coalescing doesn't work together

since write coalescing as implemented in Go can only work on connections
directly implemented in the net package of the Go stdlib. No wrappers
providing framing like TLS are allowed. Using anything else but a pure
TCP connection will lead to delayed writing of small packets instead of
write coalescing via writev.

This simply disables coalescing transparently in that case to avoid
unwanted delays.

I would have loved to return an error instead, but write coalescing is
on by default and I didn't want to suddenly return errors with the defaults.