policies: tokenAwareHostPolicy use tokens from "local" hosts only
This adds `t.fallback.IsLocal()` check to tokenAwareHostPolicy::AddHost.
Otherwise the tokenRing uses token ranges from all DCs merging them to one.
This results in wrong host selection in Pick method, primaryEndpoint
may be assigned to a host in a remote DCs. If using DCAwareRoundRobinPolicy
the host is rejected and we are falling back to a random host in a local DC.

Signed-off-by: Michał Matczuk <michal@scylladb.com>