Updates to reduce data copies and enable streamlined processing
This commit includes changes to previously defined functions to
streamline frame operations. More specificaly, it adds the raw header
bytes into the exported FrameHeader to simply copy the data w/o having
to write them out using a framer. It also updates ReadFrame to return
a copy of the full frame bytes in one go so that can be used directly
to pass them to back end etc.