Fix the round-robin policy to actually return each node.
Added a test which fails: creating two iterators and iterating in
order ABAB would return two copies of node-1 for A, two copies of
node-2 for B. This was pretty useless, since each iterator shouldn't
return the same node twice, and should return all nodes eventually.
The bug was that the index used to return the next node was global
and shared between all iterators!

The code change causes the new test to pass.