Only flag.Parse() if go version != go 1.13
This is semi fragile, but so is the way these tests are setup.

According to the go1.13 release notes:

Testing flags are now registered in the new Init function, which is invoked by the generated main function for the test. As a result, testing flags are now only registered when running a test binary, and packages that call flag.Parse during package initialization may cause tests to fail.