Fix tests in Go 1.13
The flag.Parse() in common-test.go init() causes errors to be raised
since package testing no longer registers it's flags in its own init(),
but instead provides testing.Init().

One option to fix this is to call testing.Init() just before calling
flag.Parse(). However, testing.Init() is not available in older Go
releases so we'd need to create separate versions of init() based on
build tags.

Instead, we can remove the call to flag.Parse() altogether as it was
only used to set clusterHosts variable and we defer parsing of
clusterHosts to a time when tests are executed. This version of code
runs both on Go 1.12 and Go 1.13.