session: bulk add hosts to TokenAwarePolicy on init
Adding a new host to the TokenAwarePolicy requires recomputing the
replica map, previously we would add each host one at a time which
requires regenerating the replica map for each host and discarding the
old one. Fix this to only calcaulte it once during init after adding all
the hosts in a bulk operation.