Fix unmarshaling to unsigned integers (#1360)
There were a few inconsistencies with unmarshaling to unsigned integers:

- negative values were not range checked for primitive uint types
- negative values were not allowed at all for named uint types

This commit restricts the range of accepted values so that we marshal
the same value as we unmarshal. We accept negative values for a type
only if the width of cql type is smaller or equal than the width of Go
type. This ensures that unmarshaling is bijection from CQL to values
to Go values, without the check the following situation can happen:

- we unmarshal CQL smallint 0xffff (-1) to Go uint8 0xff
- we unmarshal CQL smallint 0x00ff (255) to Go uint8 0xff
- we marshal uint8 0xff to CQL smallint 0x00ff (255)

Therefore smallint value -1 would be turned into 255 during round-trip.

We also need to apply the same logic consistently regardless of whether
we are unmarshaling into a native or named type.