Change WithTimestamp to accept time.Time
With just timestamp int64 parameter, it's possible to mistake the unit
used for the timestamp. The default timestamp is in microseconds since
epoch[1].

Using time.Time is less ambiguous and Go users are more likely to have
a time.Time value than a timestamp in microseconds since epoch.

BREAKING CHANGE: This commit changes type of a parameter in public
methods, and as such might break some users. If you have a timestamp
in microseconds since epoch and not a time.Time value, you can
convert the timestamp to time.Time with

   time.Unix(0, timestampMicro * 1e3)

[1] https://github.com/datastax/native-protocol/blob/1acbf1f40d5dfeb2a50de765bc22f12d5241ea1c/src/main/resources/native_protocol_v4.spec#L338-L343