remove iter.Scan
iter.Scan provides a poor interface to qerying rows, it requires
callers to allocate a fresh value to scan into and a value for the
initial call. This makes errors not obvious and easy to make. The
scanner interface provides a simpler and more obvious interface to
extract results from queries.