Init host pools in parallel.
Creating a session can take a long time if your cluster is big and
there's latency between you and the cluster. We are seeing it take
~10s to create a session to a 45 node cluster with ~30ms of latency.

This is somewhat related to #1369 and #1429, but this change tries to
do the simplest thing to speed up initial connection without changing
any behavior.