Connection Dial should be to connect address not hostname
Throughout gocql, this connection is mapped to the ConnectAddress which
translates to an IP. It is expected that a connection created this far
down the stack should be specific to the IP that was requested rather
than any hostname (if one is set).

In some cases (mainly when creating control connections), the hostname
is populated in the HostInfo struct with the user supplied hosts. If
these map to a DNS name with multiple entries, the connection will be
established to any of those IPs rather than specifically to the IP we
want to connect to.