Add package-level documentation (#1513)
The examples are regular godoc/go test examples.
Because go examples don't support marking the example as skipped,
the // Output line before output is intentionally omitted.
This way, the examples are compiled by go test, but not executed (as
that would fail because we need to access the database).
When someone wants to execute the example, adding the Output line
will make it executable.

Co-authored-by: Matouš Dzivjak <matousdzivjak@gmail.com>