Don't ignore TLS config's InsecureSkipVerify
tls.Config by default verifies certificates, gocql.SslOptions does
not. If one only provided the tls.Config with InsecureSkipVerify=false,
callers expect that the host will be verified, but gocql resets the
InsecureSkipVerify to true.

It's safer to explicitly disable host verification than to explicitly
enable it, so if the tls.Config is provided, let's honor it's
security settings.

If a TLS config with InsecureSkipVerify=true is provided at the same
time as EnableHostVerification=true is provided, this is a conflict
in settings. We could either return an error or fall back to verify
the host. We chose to verify the host.

The issue is in gocql codebase since commit
6495810decf1d4db412640a0f5667d4ffac62a8f, when the tls.Config was
embedded to SslOptions struct.