Fix TLS host verification when ServerName is not provided in TLS config
Enabling host verification in SslOptions fails if ServerName is
not set in the configured *tls.Config with:

gocql: unable to create session: control: unable to connect to initial hosts: tls: either ServerName or InsecureSkipVerify must be specified in the tls.Config

Before https://github.com/gocql/gocql/pull/1368 we used
tls.DialWithDialer that sets the ServerName in the tls.Config to
the host part of the connect address if ServerName was not explicitly
provided. After that MR, we wrap the connection returned by the dialer
ourselves and the ServerName is not set.

This commit restores the original behavior to set the ServerName
to host part of connect address if it is not provided explicitly
in the TLS config.

Thanks Suhail Patel for reporting the issue:
https://github.com/gocql/gocql/pull/1493