tuple_test: check unmarshaled results in nested collection test
The TestTuple_NestedCollection is supposed to test marshaling and
unmarshaling of tuples nested in collections. It first writes a row and
then reads it, but it does not check that the unmarshaled value is the
same as the one used for writing. This commit adds this check to this
test.