fix deserialization of nested tuples with nullable elements
Gocql supports deserialization of collections with tuples that hold
their elements as pointers - for example, list<tuple<string, string>>
can be deserialized into [][2]*string. Currently, the pointers in the
tuple are always set to a non-nil value, even if the tuple element was
NULL. This is inconsistent with non-nested column, for which gocql
allows to check if the column was NULL.

This patch fixes that, and now pointers in nested tuples are set to nil
if the corresponding column was NULL.

This patch also updates the TestMarshalTuple test, which expected the
old behavior.