fix unmarshaling of last empty/nil field in tuple
Tuples are serialized as a sequence of pairs: length (32bit signed
number) + contents. If a tuple element is NULL, its length is set to -1
and it has empty contents.

However, if a tuple itself is NULL, the whole tuple is represented by
empty bytes. This special case is handled in the following way: when
trying to deserialize a tuple element and the length of remaining,
unparsed bytes not greater than 4, the tuple element is treated as if
it were NULL.

However, this check is incorrect - if the last element of the tuple is
represented as length = 0 and empty contents, then the check will cause
the driver to interpret it as NULL.

For example, for type tuple<int, text>, values (1, "") and (1, NULL)
will both be deserialized into (1, NULL).

This patch fixes this problem by adjusting the check - a tuple element
is considered NULL if there are _less_ than 4 bytes remaining.