Make cluster defaults publicly accessable
Scenario:
We wrap gocql in some of our own libraries. We allow callers to override some of the ClusterConfig default values. 
However, we also want to have our help messages etc print what the default values are. 
We can copy/paste, but would prefer to just inherit the `gocql` current defaults, whatever they happen to be.

So exposed them as constants.